Add keyword matching for drun searches.
There's starting to be a fair amount of duplicated code between the drun
and normal run modes. At some point in the future, it's likely to be
worth combining them, such that there's only a single
`struct search_item` or similar, and just handle the different modes at
the start and end of execution.

+++ src/string_vec.h
@@ -7,7 +7,7 @@
 
 struct scored_string {
 	char *string;
-	int8_t search_score;
+	int32_t search_score;
 	int32_t history_score;
 };
